+++ src/passes/deferedPass.cpp
@@ -159,7 +159,7 @@ void DeferredPass::updateLights(Scene* scene) const
 		for (auto& [handle, light] : lights)
 		{
 			lightData[i] = light->getLightData();
-			lightData[i].objectID = static_cast<uint32_t>(static_cast<int32_t>(handle));
+			lightData[i].objectID = static_cast<float>(static_cast<int32_t>(handle));
 			++i;
 		}
 		m_context->Unmap(m_lightsBuffer.Get(), 0);
