+++ src/shaders/colorPicker.hlsl
@@ -8,7 +8,7 @@ cbuffer cbPickLocation : register(b0)
 
 Texture2D<float> gBufferObjectID : register(t0);
 
-RWStructuredBuffer<uint> pickedColor : register(u0);
+RWStructuredBuffer<float> pickedColor : register(u0);
 
 [numthreads(1, 1, 1)]
 void CS(uint3 DTid : SV_DISPATCHTHREADID)
